+++ pkg/eventmonitor/consumers/testutil/testutil.go
@@ -27,7 +27,8 @@ func NewTestProcessConsumer(t *testing.T) *consumers.ProcessConsumer {
 	var pc *consumers.ProcessConsumer
 	eventtestutil.StartEventMonitor(t, func(t *testing.T, evm *eventmonitor.EventMonitor) {
 		var err error
-		pc, err = consumers.NewProcessConsumer("test", defaultChanSize, evm)
+		opts := consumers.ProcessConsumerOptions{ID: "test", ChanSize: defaultChanSize}
+		pc, err = consumers.NewProcessConsumer(opts, evm)
 		require.NoError(t, err, "failed to create process consumer")
 	})
